% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\docType{data}
\name{colours_cues}
\alias{colours_cues}
\title{Colour Code Cues Used in Commonality Analysis}
\format{
An object of class \code{character} of length 9.
}
\usage{
colours_cues
}
\description{
`r lifecycle::badge("stable")`

Colours used for cues in Commonality Analysis for Valence/Arousal. Chosen for
accessible contrast.
}
\keyword{datasets}
