% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\docType{data}
\name{colours_mode}
\alias{colours_mode}
\title{Colour Code Nominal Mode}
\format{
An object of class \code{character} of length 7.
}
\usage{
colours_mode
}
\description{
`r lifecycle::badge("stable")`

Colours used in the Emotional Piano Project to distinguish Major and Minor
modes. Includes three different conventions for mode labels for flexibility.
}
\keyword{datasets}
