% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\docType{data}
\name{colours_tool}
\alias{colours_tool}
\title{Colour Code Music Content Analysis Tools}
\format{
An object of class \code{character} of length 3.
}
\usage{
colours_tool
}
\description{
`r lifecycle::badge("stable")`

Colours used for popular tools analyzed by the MAPLE Lab. Based on logos, but
modified for accessible contrast.
}
\keyword{datasets}
