% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relative_variation.R
\name{relative_variation}
\alias{relative_variation}
\title{Calculate Relative Variation}
\usage{
relative_variation(df, grouping_col, value_col, variability_metric = sd)
}
\arguments{
\item{df}{Dataframe containing the data to be analyzed}

\item{grouping_col}{Column name in `df` that defines the groups for which
variation is calculated. Each group consists of multiple versions (subgroups)}

\item{value_col}{Column name in `df` that contains the numerical values
whose variability is being measured}

\item{variability_metric}{Function to compute variability. Defaults to `sd`
(standard deviation), but other functions (e.g., `var`) can be used to
customize the measure of variation}
}
\value{
Dataframe containing the computed relative variation ratio for
each subgroup (`version`) within specified grouping. The output includes:
- `grouping_col`: The grouping identifier
- `var_version`: The variability (e.g., standard deviation) of the subgroup
- `var_total`: The total variability of the entire group
- `ratio`: The computed ratio of subgroup variation to total variation
}
\description{
`r lifecycle::badge("stable")`

Calculates the relative variation of subgroups (versions) within a larger
grouping by calculating the ratio of their variability to the total
variability of the group. This is useful for assessing the consistency of
variation within different subgroups compared to the
overall group variation.
}
\examples{
# Typical usage.
relative_variation(
  df = mtcars,
  grouping_col = cyl,
  value_col = mpg
)

# Use with dplyr:group_by().
library(dplyr)
mtcars |>
group_by(gear) |>
relative_variation(
  grouping_col = cyl,
  value_col = mpg
)

}
