% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominance-analysis.R
\name{run_dominance}
\alias{run_dominance}
\title{Dominance Analysis}
\usage{
run_dominance(formula, data, type = "general")
}
\arguments{
\item{formula}{Formula for linear regression}

\item{data}{Dataframe containing values to calculate dominance from}

\item{type}{Character. Either "general", "conditional" or "complete".}
}
\value{
Dominance values according to type argument. See details.
}
\description{
`r lifecycle::badge('stable')`
Convenience function to extract outputs from yhat::dominance function
}
\details{
This function uses arguments for formula and data to calculate dominance
values from a linear regression. When type == "general", returns a vector
with general dominance values for each DV. When type == "conditional",
returns a matrix with dominance values with columns corresponding to
the DV and rows corresponding to the number of variables in configuration.
When type == "complete", returns complete table for dominance. See
\href{this vignette}{https://rdrr.io/cran/domir/f/vignettes/domir_basics.Rmd}
for more information on interpreting dominance analysis outputs.
}
\examples{
run_dominance(mpg~cyl+hp, data = mtcars, type = "general")
run_dominance(mpg~cyl+hp, data = mtcars, type = "conditional")
run_dominance(mpg~cyl+hp, data = mtcars, type = "complete")
}
\seealso{
[run_dominance_boot()]
}
