% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{translate_filename}
\alias{translate_filename}
\title{Translate filenames into Emotional Piano Project UID}
\usage{
translate_filename(filenames)
}
\arguments{
\item{filename}{The filename to be translated. Expected format uses at least
5 underscores as delimiter, and must include the Composer, Set Number,
Performer, Chroma, Mode, and albumID. Non-conventional files can include
'Book' and Piece Number in the place of Chroma and Mode respectively. Can
include a track number (from full track files) at the beggining, and any
number of additional delimiters at the end (for instance, from manipulated
files). Can include the file extension. Note that Chroma may sometimes
include letters in case of multiple same pieceID in same set.}
}
\value{
A tibble with 3 columns:
\describe{
  \item{setCode}{(character) Unique identifier for the 'set' or work
  (e.g., Chopin Op.28)}
  \item{pieceID}{(character) Unique identifier for the specific piece. May
  include an alternate letter.}
  \item{albumID}{(character) Unique identifier for the album.}
}
}
\description{
Translates a vector of filenames into an Emotional Piano Project UID
containing a setCode, pieceID, and albumID. The function parses the provided
filenames, which is expected to follow a specific naming convention, and
extracts the unique identifiers.
}
\examples{
# Typical usage with single filename: returns a tibble.
translate_filename("Bach_1_Newman_11_minor_bachNewman2001piano.wav")
translate_filename("Debussy_1_Arrau_1_4_debussyArrau1991.wav")
translate_filename("Alkan_1_Martin_0a_Major_alkanMartin2006.wav")

# Less common usage: extract specific elements of UID indepedendantly.
translate_filename("Alkan_1_Martin_0a_Major_alkanMartin2006.wav")$setCode
translate_filename("Alkan_1_Martin_0a_Major_alkanMartin2006.wav")$pieceID
translate_filename("Alkan_1_Martin_0a_Major_alkanMartin2006.wav")$albumID

# Mutating the UIDs from filenames in a tibble.
tibble::tibble(
  filename = c(
    "Bach_1_Newman_11_minor_bachNewman2001piano.wav",
    "Debussy_1_Arrau_1_4_debussyArrau1991.wav",
    "Alkan_1_Martin_0a_Major_alkanMartin2006.wav"
  )
) |>
  dplyr::mutate(translate_filename(filename))

}
