% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominance-analysis.R
\name{run_dominance_boot}
\alias{run_dominance_boot}
\title{Dominance analysis with bootstrap percentile estimates}
\usage{
run_dominance_boot(formula, data, type = "general", n_replications = 100)
}
\arguments{
\item{formula}{Formula for linear regression}

\item{data}{Dataframe containing values to calculate dominance from}

\item{type}{Character. Either "general" or "conditional"}

\item{n_replications}{Number of bootstrap replications to perform}
}
\description{
`r lifecycle::badge('experimental')`
Calculate general or conditional dominance with bootstrap confidence
intervals
}
\details{
This function uses a formula, data syntax to calculate dominance
values along with bootstrap CIs from a linear regression. 
When type == "general", returns a vector with general dominance values for 
each DV. When type == "conditional", returns a matrix with dominance values
with columns corresponding to the DV and rows corresponding to the number of
variables in configuration.
See \href{this vignette}{https://rdrr.io/cran/domir/f/vignettes/domir_basics.Rmd} 
for more information on interpreting dominance analysis outputs.
}
\examples{
run_dominance_boot(mpg~cyl+hp, data = mtcars, type = "general")
run_dominance_boot(mpg~cyl+hp, data = mtcars, type = "conditional")
}
